// src/app/models/refs/enemy.ts
export enum EnemyRef {
  AeroCombatant = 'AeroCombatant',
  AirBuster = 'AirBuster',
  Aps = 'Aps',
  ArkDragon = 'ArkDragon',
  BloodTaste = 'BloodTaste',
  Bottomswell = 'Bottomswell',
  Capparwire = 'Capparwire',
  Castanets = 'Castanets',
  Crawler = 'Crawler',
  Deenglow = 'Deenglow',
  Eligor = 'Eligor',
  Elfadunk = 'Elfadunk',
  FirstRay = 'FirstRay',
  Formula = 'Formula',
  Grunt = 'Grunt',
  GuardHound = 'GuardHound',
  GuardScorpion = 'GuardScorpion',
  HedgehogPie = 'HedgehogPie',
  HeliGunner = 'HeliGunner',
  HellHouse = 'HellHouse',
  HellRider = 'HellRider',
  HundredGunner = 'HundredGunner',
  KalmFang = 'KalmFang',
  Levrikon = 'Levrikon',
  Madouge = 'Madouge',
  Mandragora = 'Mandragora',
  MidgarZolom = 'MidgarZolom',
  MightyGrunt = 'MightyGrunt',
  MotorBall = 'MotorBall',
  MP = 'MP',
  MysteryNinja = 'MysteryNinja',
  ProtoMachinegun = 'ProtoMachinegun',
  Reno = 'Reno',
  Rufus = 'Rufus',
  SampleH0512 = 'SampleH0512',
  Smogger = 'Smogger',
  Soldier3rd = 'Soldier3rd',
  SpecialCombatant = 'SpecialCombatant',
  Sweeper = 'Sweeper',
  SwordDance = 'SwordDance',
  VargidPolice = 'VargidPolice',
  Vice = 'Vice',
  WholeEater = 'WholeEater',
  Zemzelett = 'Zemzelett',
  Zenene = 'Zenene',
}